added footer steps
